+++ .github/workflows/frontend.yml
@@ -1,0 +1,56 @@
+name: Frontend CI
+
+on:
+  push:
+    branches: ['main']
+  pull_request:
+    branches: ['main']
+
+jobs:
+  lint-format-build:
+    runs-on: ubuntu-latest
+    permissions:
+      contents: write 
+
+    defaults:
+      run:
+        working-directory: ./frontend
+
+    steps:
+      - name: Checkout code
+        uses: actions/checkout@v4
+        with:
+          # On pull requests, checkout the head ref to be able to push changes
+          ref: ${{ github.event.pull_request.head.ref }}
+
+      - name: Setup Node.js
+        uses: actions/setup-node@v4
+        with:
+          node-version: '20'
+          cache: 'npm'
+          cache-dependency-path: frontend/package-lock.json
+
+      - name: Install dependencies
+        run: npm install
+
+      - name: Check formatting
+        id: format-check
+        run: npm run format:check
+        continue-on-error: true
+
+      - name: Auto-format code if check fails
+        if: steps.format-check.outcome == 'failure'
+        run: npm run format
+
+      - name: Commit and push formatting changes
+        if: steps.format-check.outcome == 'failure'
+        run: |
+          git config --global user.name 'github-actions[bot]'
+          git config --global user.email 'github-actions[bot]@users.noreply.github.com'
+          git add frontend
+          git commit -m "Prettier formatted the code"
+          git push
+        working-directory: .
+
+      - name: Build project
+        run: npm run build
