+++ frontend/package.json
@@ -11,7 +11,7 @@
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
     "react-quizlet-flashcard": "^3.0.0",
-    "react-router-dom": "^7.5.1",
+    "react-router-dom": "^7.6.3",
     "react-scripts": "^5.0.1",
     "react-slick": "^0.30.3",
     "slick-carousel": "^1.8.1"
