+++ frontend/src/assets/technotes_logo.jpg
