+++ frontend/src/components/Notes.jsx
@@ -1,19 +1,16 @@
-import React from 'react';
-import { notes } from '../utils/shm';
+import { notes } from '../utils/notes';
 
-const Sim = () => {
+const Notes = () => {
   return (
-    <div className="flex flex-col items-center gap-8">
+    <div className="flex h-screen flex-col items-center gap-8">
       {notes.map((note) => (
         <div
           key={note.id}
           className="bg-white rounded-xl shadow-lg p-6 hover:shadow-xl transition-shadow w-full max-w-2xl"
         >
-          {/* Τίτλος & Κατηγορία */}
           <h3 className="text-xl font-semibold mb-2 text-gray-800">{note.title}</h3>
           <p className="text-sm text-pink-600 mb-3">{note.subject}</p>
 
-          {/* Προεπισκόπηση PDF */}
           {note.pdfUrl && (
             <div className="mb-4">
               <iframe
@@ -26,26 +23,24 @@ const Sim = () => {
             </div>
           )}
 
-          {/* Κουμπιά PDF */}
-          <div className="flex gap-3 mb-4">
+          <div className="flex flex-col items-center gap-3 mb-4">
             <a
               href={note.pdfUrl}
               target="_blank"
               rel="noopener noreferrer"
-              className="text-sm underline text-blue-600 hover:text-blue-800"
+              className="w-full max-w-xs text-center py-2 px-4 rounded-lg transition-colors bg-blue-500 text-white hover:bg-blue-600"
             >
               Προβολή σε νέο παράθυρο
             </a>
             <a
               href={note.pdfUrl}
               download
-              className="text-sm bg-gray-100 px-3 py-1 rounded hover:bg-gray-200"
+              className="w-full max-w-xs text-center py-2 px-4 rounded-lg transition-colors bg-gray-200 text-gray-800 hover:bg-gray-300"
             >
-              Κατέβασε PDF
+              Λήψη PDF
             </a>
           </div>
 
-          {/* Τιμή & κουμπί αγοράς */}
           <div className="flex justify-between items-center">
             <span className="text-2xl font-bold text-pink-600">€{note.price}</span>
             <button className="bg-pink-500 text-white px-4 py-2 rounded-lg hover:bg-pink-600 transition-colors">
@@ -58,4 +53,4 @@ const Sim = () => {
   );
 };
 
-export default Sim;
+export default Notes;
